Move scripts to scenarios folder
