+++ scripts/k6/libs/auth.js
