+++ scripts/k6/scenarios/create-holders.js
@@ -4,19 +4,18 @@
 
 import { sleep, check } from "k6";
 import { SharedArray } from "k6/data";
-import { getBearerToken, getGovernanceBearerToken } from "./auth.js";
+import { getBearerToken, getGovernanceBearerToken } from "../libs/auth.js";
 import { Trend, Counter } from "k6/metrics";
 import file from "k6/x/file";
+import { createSchemaIfNotExists } from "../libs/schemaUtils.js";
+import { createIssuerIfNotExists } from '../libs/issuerUtils.js';
 import {
   createTenant,
   getWalletIdByWalletName,
-  getAccessTokenByWalletId,
   deleteTenant,
-  createIssuerTenant,
   createCredentialDefinition,
   getCredentialDefinitionId,
-  createSchema
-} from "./tenant.js";
+} from "../libs/functions.js";
 
 const vus = parseInt(__ENV.VUS, 10);
 const iterations = parseInt(__ENV.ITERATIONS, 10);
@@ -41,6 +40,7 @@ export let options = {
     "http_req_duration{scenario:default}": ["max>=0"],
     "http_reqs{scenario:default}": ["count >= 0"],
     "iteration_duration{scenario:default}": ["max>=0"],
+    "checks": ["rate==1"],
     // 'specific_function_reqs{my_custom_tag:specific_function}': ['count>=0'],
     // 'specific_function_reqs{scenario:default}': ['count>=0'],
   },
@@ -80,37 +80,16 @@ export function setup() {
   for (let i = 0; i < numIssuers; i++) {
     const walletName = `${issuerPrefix}_${i}`;
     const credDefTag = walletName;
-    // const schemaId = __ENV.SCHEMA_ID;
 
-    let issuerAccessToken;
-    let issuerWalletId;
-
-    issuerWalletId = getWalletIdByWalletName(bearerToken, walletName);
-    if (issuerWalletId !== null) {
-      // Retrieve the access token using the wallet ID
-      issuerAccessToken = getAccessTokenByWalletId(bearerToken, issuerWalletId);
-      if (typeof issuerAccessToken === "string") {
-        // Access token retrieved successfully
-        console.log(`Access token retrieved for wallet ID ${issuerWalletId}`);
-      } else {
-        console.error(`Failed to retrieve access token for wallet ID ${issuerWalletId}`);
-        console.error(`Response body: ${issuerAccessToken}`);
-        continue;
-      }
-    } else {
-      try {
-        const createIssuerTenantResponse = createIssuerTenant(bearerToken, walletName);
-        check(createIssuerTenantResponse, {
-          "Issuer tenant created successfully": (r) => r.status === 200
-        });
-        const tenantData = JSON.parse(createIssuerTenantResponse.body);
-        issuerWalletId = tenantData.wallet_id;
-        issuerAccessToken = tenantData.access_token;
-      } catch (error) {
-        console.error(`Error creating issuer tenant for ${walletName}:`, error);
-        continue;
-      }
+    const issuerData = createIssuerIfNotExists(bearerToken, walletName);
+    check(issuerData, {
+      "Issuer data retrieved successfully": (data) => data !== null && data !== undefined
+    });
+    if (!issuerData) {
+      console.error(`Failed to create or retrieve issuer for ${walletName}`);
+      continue;
     }
+    const { issuerWalletId, issuerAccessToken } = issuerData;
 
     const credentialDefinitionId = getCredentialDefinitionId(bearerToken, issuerAccessToken, credDefTag);
     if (credentialDefinitionId) {
@@ -126,13 +105,10 @@ export function setup() {
       // console.error(`Response body: ${credentialDefinitionId.body}`);
     }
 
-    const createSchemaResponse = createSchema(governanceBearerToken, schemaName, schemaVersion);
-    check(createSchemaResponse, {
-      "Schema created successfully (or existed already)": (r) => r.status === 200
+    const schemaId = createSchemaIfNotExists(governanceBearerToken, schemaName, schemaVersion);
+    check(schemaId, {
+      "Schema ID is not null": (id) => id !== null && id !== undefined
     });
-    const { id: schemaId } = JSON.parse(createSchemaResponse.body);
-
-    console.log(`Schema ID: ${schemaId}`);
 
     const createCredentialDefinitionResponse = createCredentialDefinition(bearerToken, issuerAccessToken, credDefTag, schemaId);
     check(createCredentialDefinitionResponse, {
