+++ scripts/k6/scenarios/create-invitation.js
@@ -2,34 +2,27 @@
 /* eslint no-undef: "error" */
 /* eslint no-console: ["error", { allow: ["warn", "error", "log"] }] */
 
-import { check, sleep } from "k6";
+import { check } from "k6";
 import { SharedArray } from "k6/data";
-import { getBearerToken } from "./auth.js";
+import { getBearerToken } from "../libs/auth.js";
 import { Trend, Counter } from "k6/metrics";
+import file from "k6/x/file";
+import sleep from "k6";
 import {
   getWalletIdByWalletName,
   getAccessTokenByWalletId,
   deleteTenant,
   createIssuerTenant,
-  createCredential,
-  acceptCredential,
+  createInvitation,
+  acceptInvitation,
   createCredentialDefinition,
-  getCredentialIdByThreadId,
-  waitForSSEEvent,
-  getCredentialDefinitionId,
-  getProofIdCredentials,
-  sendProofRequest,
-  waitForSSEEventReceived,
-  getProofIdByThreadId,
-  acceptProofRequest,
-  waitForSSEProofDone,
-  getProof,
-} from "./tenant.js";
+  waitForSSEEventConnection,
+  getCredentialDefinitionId
+} from "../libs/functions.js";
 
 const vus = parseInt(__ENV.VUS, 10);
 const iterations = parseInt(__ENV.ITERATIONS, 10);
 const issuerPrefix = __ENV.ISSUER_PREFIX;
-
 
 export let options = {
   scenarios: {
@@ -40,40 +33,43 @@ export let options = {
       maxDuration: "24h",
     },
   },
-  setupTimeout: "180s", // Increase the setup timeout to 120 seconds
-  teardownTimeout: "180s", // Increase the teardown timeout to 120 seconds
+  setupTimeout: "120s", // Increase the setup timeout to 120 seconds
+  teardownTimeout: "120s", // Increase the teardown timeout to 120 seconds
   maxRedirects: 4,
   thresholds: { //https://community.grafana.com/t/ignore-http-calls-made-in-setup-or-teardown-in-results/97260/2
     "http_req_duration{scenario:default}": ["max>=0"],
     "http_reqs{scenario:default}": ["count >= 0"],
+    "http_reqs{my_custom_tag:specific_function}": ["count>=0"],
     "iteration_duration{scenario:default}": ["max>=0"],
-    // 'specific_function_reqs{my_custom_tag:specific_function}': ['count>=0'],
-    // 'specific_function_reqs{scenario:default}': ['count>=0'],
+    "checks": ["rate==1"],
+    // 'test_function_reqs{my_custom_tag:specific_function}': ['count>=0'],
+    // 'test_function_reqs{scenario:default}': ['count>=0'],
+    // 'custom_duration{step:getAccessTokenByWalletId}': ['avg>=0'],
   },
   tags: {
     test_run_id: "phased-issuance",
-    test_phase: "create-proofs",
+    test_phase: "create-invitation",
   },
 };
 
-const inputFilepath = "output/create-invitation.json";
-const data = open(inputFilepath, "r");
-
-// const specificFunctionReqs = new Counter('specific_function_reqs');
 const testFunctionReqs = new Counter("test_function_reqs");
-// const mainIterationDuration = new Trend('main_iteration_duration');
+const mainIterationDuration = new Trend("main_iteration_duration");
+
+const inputFilepath = "../output/create-holders.json";
+const data = open(inputFilepath, "r");
+const outputFilepath = "output/create-invitation.json";
 
 // Seed data: Generating a list of options.iterations unique wallet names
-// const wallets = new SharedArray('wallets', function() {
-//   const walletsArray = [];
-//   for (let i = 0; i < options.iterations; i++) {
-//     walletsArray.push({
-//       wallet_label: `xk6 holder ${i}`,
-//       wallet_name: `xk6_wallet_${i}`
-//     });
-//   }
-//   return walletsArray;
-// });
+const wallets = new SharedArray("wallets", function() {
+  const walletsArray = [];
+  for (let i = 0; i < options.iterations; i++) {
+    walletsArray.push({
+      wallet_label: `xxkk6 holder ${i}`,
+      wallet_name: `xxkk6_wallet_${i}`
+    });
+  }
+  return walletsArray;
+});
 
 const numIssuers = 1;
 let issuers = [];
@@ -89,6 +85,8 @@ export function setup() {
   //   console.log(`Processing wallet ID: ${holderData.wallet_id}`);
   //   // Your test logic here, e.g., make HTTP requests using the holderData
   // });
+
+  file.writeString(outputFilepath, "");
 
   for (let i = 0; i < numIssuers; i++) {
     const walletName = `${issuerPrefix}_${i}`;
@@ -126,7 +124,7 @@ export function setup() {
 
     const credentialDefinitionId = getCredentialDefinitionId(bearerToken, issuerAccessToken, credDefTag);
     if (credentialDefinitionId) {
-      console.warn(`Credential definition already exists for issuer ${walletName} - Skipping creation`);
+      console.log(`Credential definition already exists for issuer ${walletName} - Skipping creation`);
       issuers.push({
         walletId: issuerWalletId,
         accessToken: issuerAccessToken,
@@ -166,47 +164,61 @@ function getWalletIndex(vu, iter) {
   return walletIndex;
 }
 
-//random number between 0 and 100 (including 0 and 100 as options)
-function getRandomInt() {
-  return Math.floor(Math.random() * 101);
-}
+const vuStartTimes = {};
+const vuEndTimes = {};
 
 export default function(data) {
-  // const start = Date.now();
+  if (__ITER === 0) {
+    vuStartTimes[__VU] = Date.now();
+  }
+  const start = Date.now();
   const bearerToken = data.bearerToken;
   const issuers = data.issuers;
-  const holders = data.holders;
   const walletIndex = getWalletIndex(__VU, __ITER + 1); // __ITER starts from 0, adding 1 to align with the logic
+
+  const holders = data.holders;
   const wallet = holders[walletIndex];
+
 
   const issuerIndex = __ITER % numIssuers;
   const issuer = issuers[issuerIndex];
 
-  // console.log(`isser.accessToken: ${issuer.accessToken}`);
-  // console.log(`issuer.credentialDefinitionId: ${issuer.credentialDefinitionId}`);
-  // console.log(`wallet.issuer_connection_id: ${wallet.issuer_connection_id}`);
-  // const sendProofRequestResponse = sendProofRequest(issuer.accessToken, wallet.issuer_connection_id);
-  let sendProofRequestResponse;
-  try {
-    sendProofRequestResponse = sendProofRequest(issuer.accessToken, wallet.issuer_connection_id);
-  } catch (error) {
-      // console.error(`Error creating credential: ${error.message}`);
-      sendProofRequestResponse = { status: 500, response: error.message };
-  }
-  check(sendProofRequestResponse, {
-    "Proof request sent successfully": (r) => {
+  // const holderWalletId = getWalletIdByWalletName(bearerToken, wallet.wallet_name);
+  // check(holderWalletId, {
+  //   "Holder wallet ID is not null": (r) => r !== null
+  // });
+
+  // const holderAccessToken = getAccessTokenByWalletId(bearerToken, holderWalletId);
+  // check(holderAccessToken, {
+  //   "Holder access token is not null": (r) => r !== null
+  // });
+
+  const createInvitationResponse = createInvitation(bearerToken, issuer.accessToken);
+  check(createInvitationResponse, {
+    "Invitation created successfully": (r) => {
       if (r.status !== 200) {
-        throw new Error(`Unexpected response while sending proof request: ${r.response}`);
+        throw new Error(`Unexpected response status while create invitation: ${r.status}`);
+      }
+      return true;
+    }
+  });
+  const { invitation: invitationObj, connection_id: issuerConnectionId } = JSON.parse(createInvitationResponse.body);
+
+  const acceptInvitationResponse = acceptInvitation(wallet.access_token, invitationObj);
+  check(acceptInvitationResponse, {
+    "Invitation accepted successfully": (r) => {
+      if (r.status !== 200) {
+        throw new Error(`Unexpected response while accepting invitation: ${r.response}`);
       }
       return true;
     }
   });
 
-  const { thread_id: threadId } = JSON.parse(sendProofRequestResponse.body);
+  const { connection_id: holderInvitationConnectionId } = JSON.parse(acceptInvitationResponse.body);
 
-  const waitForSSEEventReceivedResponse = waitForSSEEventReceived(wallet.access_token, wallet.wallet_id, threadId);
-  check(waitForSSEEventReceivedResponse, {
-    "SSE Event received successfully: request-recevied": (r) => {
+  const waitForSSEEventConnectionResponse = waitForSSEEventConnection(wallet.access_token, wallet.wallet_id, holderInvitationConnectionId);
+  check(waitForSSEEventConnectionResponse, {
+    "SSE Event received successfully: connection-ready": (r) => {
       if (!r) {
         throw new Error("SSE event was not received successfully");
       }
@@ -214,56 +226,33 @@ export default function(data) {
     },
   });
 
-  // TODO: return object and add check for the response
-  const proofId = getProofIdByThreadId(wallet.access_token, threadId);
-  const referent = getProofIdCredentials(wallet.access_token, proofId);
-
-  const acceptProofResponse = acceptProofRequest(wallet.access_token, proofId, referent);
-  check(acceptProofResponse, {
-    "Proof accepted successfully": (r) => {
-      if (r.status !== 200) {
-        throw new Error(`Unexpected response while accepting proof: ${r.response}`);
-      }
-      return true;
-    }
-  });
-
-  const waitForSSEProofDoneRequest = waitForSSEProofDone(issuer.accessToken, issuer.walletId, threadId);
-  check(waitForSSEProofDoneRequest, {
-    "SSE Proof Request state: done": (r) => {
-      if (!r) {
-        throw new Error("SSE proof done was not successful");
-      }
-      return true;
-    },
-  });
-
-  // const getProofResponse = getProof(issuer.accessToken, wallet.issuer_connection_id, threadId );
-  let getProofResponse;
-  try {
-    getProofResponse = getProof(issuer.accessToken, wallet.issuer_connection_id, threadId );
-  } catch (error) {
-      // console.error(`Error creating credential: ${error.message}`);
-      getProofResponse = { status: 500, response: error.message };
-  }
-  check(getProofResponse, {
-    "Proof received successfully": (r) => {
-      if (r.status !== 200) {
-        throw new Error(`Unexpected response while getting proof: ${r.response}`);
-      }
-      return true;
-    },
-  });
+  // testFunctionReqs.add(1, { my_custom_tag: 'specific_function' });
 
   testFunctionReqs.add(1);
+
+  const holderData = JSON.stringify({
+    wallet_label: wallet.wallet_label,
+    wallet_name: wallet.wallet_name,
+    wallet_id: wallet.wallet_id,
+    access_token: wallet.access_token,
+    connection_id: holderInvitationConnectionId,
+    issuer_connection_id: issuerConnectionId,
+  });
+  file.appendString(outputFilepath, holderData + "\n");
+
+  const end = Date.now();
+  const duration = end - start;
+  // console.log(`Duration for iteration ${__ITER}: ${duration} ms`);
+  mainIterationDuration.add(duration);
 }
 
+
+
 export function teardown(data) {
+  vuEndTimes[__VU] = Date.now();
   const bearerToken = data.bearerToken;
   const issuers = data.issuers;
-  const wallets = data.holders;
-
-  // console.log(__ENV.SKIP_DELETE_ISSUERS)
+  const holders = data.holders;
 
   if (__ENV.SKIP_DELETE_ISSUERS !== "true") {
     for (const issuer of issuers) {
@@ -285,7 +274,7 @@ export function teardown(data) {
   }
   // // Delete holder tenants
   if (__ENV.SKIP_DELETE_HOLDERS !== "true") {
-    for (const wallet of wallets) {
+    for (const wallet of holders) {
       const walletId =  getWalletIdByWalletName(bearerToken, wallet.wallet_name);
       const deleteHolderResponse = deleteTenant(bearerToken, walletId);
       check (deleteHolderResponse, {
